requirements.txt
Django==1.11.2
django-nested-admin==3.0.17
python-monkey-business==1.0.0
pytz==2017.2
six==1.10.0
